# chore(sql): optimize varchar like operator for short contains patterns