--- a/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
+++ b/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
@@ -123,9 +123,9 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
                         this.fieldHi += 7;
                         continue;
                     } else {
-                        long idx = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
-                        ptr += idx;
-                        this.fieldHi += idx;
+                        int firstIndex = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
+                        ptr += firstIndex;
+                        this.fieldHi += firstIndex;
                     }
                 }
 
@@ -329,9 +329,9 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
                         this.fieldHi += 7;
                         continue;
                     } else {
-                        long idx = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
-                        ptr += idx;
-                        this.fieldHi += idx;
+                        int firstIndex = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
+                        ptr += firstIndex;
+                        this.fieldHi += firstIndex;
                     }
                 }