--- a/core/src/main/java/io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java
@@ -151,7 +151,6 @@ public abstract class AbstractLikeVarcharFunctionFactory implements FunctionFact
      */
     private static class ConstContainsSwarVarcharFunction extends BooleanFunction implements UnaryFunction {
         private static final int MAX_SIZE = Long.BYTES;
-        private static final long ZERO_CHECK_MASK = SwarUtils.checkZeroByte(0);
         private final Utf8Sequence pattern; // only used in toPlan
         private final long patternMask;
         private final int patternSize;
@@ -192,14 +191,12 @@ public abstract class AbstractLikeVarcharFunctionFactory implements FunctionFact
                 long check = SwarUtils.checkZeroByte(us.longAt(i) ^ searchWord);
                 if (check != 0) {
                     // We've found a match for the first byte, slow down and check the full pattern.
-
-                    // Find the lowest 0x80 byte in the check result and calculate its index.
-                    // Such byte stands for a search match, so we start the pattern comparison with it.
-                    int start = (Long.numberOfTrailingZeros(check & ZERO_CHECK_MASK) - 7) / 8;
-                    for (int j = start; j < 8; j++) {
-                        if ((us.longAt(i + j) & patternMask) == patternWord) {
+                    for (int j = 0; j < 8; j++) {
+                        // Check if the pattern matches only for matched first bytes.
+                        if ((check & SwarUtils.ZERO_CHECK_BYTE) != 0 && (us.longAt(i + j) & patternMask) == patternWord) {
                             return true;
                         }
+                        check >>>= 8;
                     }
                 }
             }