--- a/core/src/main/java/io/questdb/std/SwarUtils.java
+++ b/core/src/main/java/io/questdb/std/SwarUtils.java
@@ -38,6 +38,9 @@ public final class SwarUtils {
 
     /**
      * Returns non-zero result in case if the input contains a zero byte.
+     * <p>
+     * Each zero byte of the input is replaced with 0x80 in the output.
+     * Each non-zero byte is replaced with zero byte (0x00).
      */
     public static long checkZeroByte(long w) {
         return ((w - 0x0101010101010101L) & ~(w) & 0x8080808080808080L);