--- a/core/src/main/java/io/questdb/std/SwarUtils.java
+++ b/core/src/main/java/io/questdb/std/SwarUtils.java
@@ -26,6 +26,8 @@ package io.questdb.std;
 
 public final class SwarUtils {
 
+    public static final long ZERO_CHECK_BYTE = 0x80;
+
     private SwarUtils() {
     }
 
@@ -39,8 +41,8 @@ public final class SwarUtils {
     /**
      * Returns non-zero result in case if the input contains a zero byte.
      * <p>
-     * Each zero byte of the input is replaced with 0x80 in the output.
-     * Each non-zero byte is replaced with zero byte (0x00).
+     * Each zero byte of the input is replaced with {@link #ZERO_CHECK_BYTE} in the output.
+     * Each non-zero byte is replaced with zero byte.
      */
     public static long checkZeroByte(long w) {
         return ((w - 0x0101010101010101L) & ~(w) & 0x8080808080808080L);