--- a/core/src/main/java/io/questdb/std/SwarUtils.java
+++ b/core/src/main/java/io/questdb/std/SwarUtils.java
@@ -41,10 +41,10 @@ public final class SwarUtils {
 
     /**
      * Returns index of lowest (LE) non-zero byte in the input number
-     * or 7 in case if the number is zero.
+     * or 8 in case if the number is zero.
      */
-    public static long indexOfFirstMarkedByte(long w) {
-        return ((((w - 1) & 0x101010101010101L) * 0x101010101010101L) >> 56) - 1;
+    public static int indexOfFirstMarkedByte(long w) {
+        return Long.numberOfTrailingZeros(w) >>> 3;
     }
 
     /**