… core/src/main/java/io/questdb/std/str/Utf8Sequence.java