Adds SWAR-based implementation for `like '%mypattern%'` operations with patterns up to 8 bytes in size. The improvement comes from first byte search done in 8-byte steps and, also, an 8-byte pattern match check.

Improvement on the ClickBench data set:
```sql
-- master 3.5s
-- patch 1s
SELECT COUNT(*) FROM hits WHERE URL LIKE '%google%';

-- master 2.2s
-- patch 0.7s
SELECT SearchPhrase, MIN(URL), MIN(Title), COUNT(*) AS c, count_distinct(UserID) FROM hits WHERE Title LIKE '%Google%' AND URL NOT LIKE '%.google.%' AND SearchPhrase IS NOT NULL GROUP BY SearchPhrase ORDER BY c DESC LIMIT 10;
```