@ideoma 2024-04-02
#### [PR Coverage check]

:heart_eyes: **pass** : 38 / 45 (84.44%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|38|45|84.44%|