@ideoma 2024-04-02
#### [PR Coverage check]

:heart_eyes: **pass** : 39 / 46 (84.78%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|37|44|84.09%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|2|2|100.00%|