@ideoma 2024-04-02
#### [PR Coverage check]

:heart_eyes: **pass** : 42 / 48 (87.50%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|40|46|86.96%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|2|2|100.00%|