@ideoma 2024-04-03
#### [PR Coverage check]

:heart_eyes: **pass** : 46 / 52 (88.46%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|44|50|88.00%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|2|2|100.00%|