@ideoma 2024-04-04
#### [PR Coverage check]

:heart_eyes: **pass** : 48 / 61 (78.69%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8Sequence.java|0|7|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|46|52|88.46%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|2|2|100.00%|