@ideoma 2024-04-10
#### [PR Coverage check]

:heart_eyes: **pass** : 48 / 55 (87.27%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8Sequence.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|46|52|88.46%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|2|2|100.00%|