@ideoma 2024-04-11
#### [PR Coverage check]

:heart_eyes: **pass** : 59 / 66 (89.39%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8Sequence.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeVarcharFunctionFactory.java|49|55|89.09%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CsvFileIndexer.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/AbstractTextLexer.java|6|6|100.00%|